refactored app layout: - moved some components into modules `navigation` and `layout` - each page now renders the navbar separately
