+++ src/modules/Layout/Head.tsx
@@ -1,16 +1,12 @@
 //* Core
-import { FC, useEffect } from "react";
+import { FC } from "react";
 import Head from "next/head";
 import { MetaData } from "@util/Config";
 
-//* Redux
-import { useAppDispatch } from "@store/hooks";
-import { clientActions } from "@store/Client/slice";
-
-type SEOProps = {
+export type SEOProps = {
   title: string;
   flipOrder?: boolean;
-  description: string;
+  description?: string;
   slug?: string;
   hideNavbar?: boolean;
 };
@@ -18,18 +14,9 @@ type SEOProps = {
 const StreamPiSEO: FC<SEOProps> = ({
   title,
   flipOrder,
-  description,
+  description = MetaData.description,
   slug,
-  hideNavbar = false,
 }) => {
-  //* REDUX
-  const dispatch = useAppDispatch();
-
-  //* Core
-  useEffect(() => {
-    dispatch(clientActions.setShowNavbar(!hideNavbar));
-  }, [hideNavbar, dispatch]);
-
   const titleString = flipOrder
     ? `${MetaData.title} ${title}`
     : `${title} ${MetaData.title}`;
