+++ src/pages/404.tsx
@@ -1,18 +1,15 @@
 import { FC, CSSProperties } from "react";
-import Head from "next/head";
 import Link from "next/link";
 import Button from "@components/Button";
 import { useRouter } from "next/router";
+import Layout from "@modules/Layout";
 
 const Custom404: FC = () => {
   const router = useRouter();
 
   const code: CSSProperties = { color: "inherit", fontWeight: "bold" };
   return (
-    <>
-      <Head>
-        <title>Resource Not Found</title>
-      </Head>
+    <Layout title="Resource Not Found">
       <div className="text-center">
         <h2 className="mt-5">Error 404</h2>
         <h1>Resource Not found</h1>
@@ -31,7 +28,7 @@ const Custom404: FC = () => {
           <Button className="mt-5">Go Home</Button>
         </Link>
       </div>
-    </>
+    </Layout>
   );
 };
