+++ src/pages/_app.tsx
@@ -14,12 +14,6 @@ import { useEffect } from "react";
 import useHashChange from "@hooks/useHashChange";
 import useInfoBanner from "@hooks/useInfoBanner";
 import { ToastContainer, Zoom } from "react-toastify";
-import Container from "react-bootstrap/Container";
-import StreamPiNavItem from "@components/Navigation/NavItem";
-import StreamPiFooter from "@components/Footer";
-import FooterColumn from "@components/Footer/Column";
-import StreamPiNavbar from "@components/Navbar";
-import ThemeSwitch from "@components/ThemeSwitch";
 import ScrollToTop from "@components/ScrollToTop";
 
 //* REDUX
@@ -50,55 +44,7 @@ function MyApp({ Component, pageProps }: AppProps) {
         position="top-center"
         enableMultiContainer
       />
-      <StreamPiNavbar>
-        <StreamPiNavItem to="/">Home</StreamPiNavItem>
-        <StreamPiNavItem to="/about">About</StreamPiNavItem>
-        <StreamPiNavItem to="/features">Features</StreamPiNavItem>
-        <StreamPiNavItem to="/contact">Contact</StreamPiNavItem>
-        <StreamPiNavItem to="/install">Install</StreamPiNavItem>
-        <StreamPiNavItem to="/troubleshooting">Troubleshooting</StreamPiNavItem>
-        <ThemeSwitch />
-      </StreamPiNavbar>
-      <Container style={{ paddingTop: "4rem" }} fluid="md">
-        <Component {...pageProps} />
-      </Container>
-      <StreamPiFooter>
-        <FooterColumn
-          header={"Links"}
-          links={[
-            {
-              name: "Documentation",
-              href: "https://documentation.stream-pi.com",
-            },
-            {
-              name: "Roadmap",
-              href: "https://github.com/orgs/stream-pi/projects/2",
-            },
-          ]}
-        />
-        <FooterColumn
-          header={"Social"}
-          links={[
-            {
-              name: "Twitter",
-              href: "https://twitter.com/stream_pi",
-            },
-          ]}
-        />
-        <FooterColumn
-          header={"Community"}
-          links={[
-            {
-              name: "Discord",
-              href: "https://discord.gg/BExqGmk",
-            },
-            {
-              name: "Matrix",
-              href: "https://matrix.to/#/!hTwUYZonUXThjkMhCD:matrix.org?via=matrix.org",
-            },
-          ]}
-        />
-      </StreamPiFooter>
+      <Component {...pageProps} />
       {/* For Site Updates - useInfoBanner */}
       <ToastContainer
         position="bottom-center"
