+++ src/pages/contact.tsx
@@ -3,20 +3,19 @@ import Container from "react-bootstrap/Container";
 import Row from "react-bootstrap/Row";
 import Col from "react-bootstrap/Col";
 import Button from "@components/Button";
-import StreamPiSEO from "@components/StreamPiSEO";
 import { ContactForm, PreFormInfo } from "@modules/Contact";
 import { useCountDown } from "@hooks/useCountDown";
+import Layout from "@modules/Layout";
 
 const StreamPiContact: FC = () => {
   const [agreed, setAgreed] = useState(false);
   const { counting, remainingSeconds } = useCountDown(8000);
 
   return (
-    <>
-      <StreamPiSEO
-        title="Contact"
-        description="Get into contact with the Stream-Pi team! Ask us a question, give us some feedback, we'd love to hear from you!"
-      />
+    <Layout
+      title="Contact"
+      description="Get into contact with the Stream-Pi team! Ask us a question, give us some feedback, we'd love to hear from you!"
+    >
       <Row className="pt-2 animate__animated animate__fadeIn">
         <Col className="text-center">
           <h3>Reach out to us</h3>
@@ -44,7 +43,7 @@ const StreamPiContact: FC = () => {
           </Container>
         </Col>
       </Row>
-    </>
+    </Layout>
   );
 };
